Adding some new commands to the Tcl directory.
debug - A very handy command for printing debugging information.

html  - A Tcl re-write of the NWS html command for printing text in HTML.

incr0 - Increment a variable.  If it doesn't exist, create it.

wrap  - Take a string and wrap it at X characters.

read_file - Read the contents of a file and return it.