Adding a new package called 'DIO' (Database Interface Objects).
This package is a set of classes built to handle SQL databases in a generic
interface.  It is still under development, but it's pretty well stable
under Postgres and pretty nice under MySQL as well.

The other package is 'DIODisplay' which is a set of classes built to display
database interfaces on the web using DIO objects to do the real back-end
work.

Documentation will follow shortly.