Add support for hidden variables.  This is so that you can have state variables that are maintained by the function outside of DIOdisplay, retained as you jump through DIOdisplay.
Add DIORowHeader to the style sheet.

Turn the search table into a navigation bar.

Added a filter method that allows you to edit the display of fields on the
fly, for example, you have a time in integer seconds since 1970 and you can
add a filter that does a "clock format" using -filter, or, for example, it
can look up an IP address from a table.

These changes were provided courtesy of Peter da Silva and Superconnect, Ltd.